pyproject.toml
[project]
name = "genshi"
version = "1.0.0"
description = "GenShi — AI-powered spreadsheet generator. Real B2B data via Netrows + LLM enrichment via AiAssist."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "dnspython>=2.8.0",
    "email-validator>=2.3.0",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "pydantic>=2.13.3",
    "python-multipart>=0.0.26",
    "sqlalchemy>=2.0.49",
    "sse-starlette>=3.3.4",
    "uvicorn[standard]>=0.45.0",
]

[project.urls]
Homepage = "https://interchained.io"
Repository = "https://github.com/interchained/genshi"
